[Feat]: 생선 에셋 수정
기존의 생선 재료들의 아이콘을 수정하고 에셋을 수정 중에 있습니다.
