+++ Assets/ScriptableObjects/FoodObjectSO/FoodDatabase.asset
@@ -152,7 +152,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: 5b76a4b668c34445a857ca2853577f09, type: 3}
   - <name>k__BackingField: Saba
     <food>k__BackingField: 5
     <tag>k__BackingField: 0
@@ -162,7 +162,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: b406c20ce7c6746d18dbdb02af535add, type: 3}
   - <name>k__BackingField: Tuna
     <food>k__BackingField: 7
     <tag>k__BackingField: 0
@@ -172,7 +172,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: a03db95ecd7dc474c94741c2af853045, type: 3}
   - <name>k__BackingField: TunaDoro
     <food>k__BackingField: 8
     <tag>k__BackingField: 0
@@ -182,7 +182,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: 1e50eccb2252449b3935b9d6be27fa58, type: 3}
   - <name>k__BackingField: Salmon
     <food>k__BackingField: 9
     <tag>k__BackingField: 0
@@ -192,7 +192,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: f721a88c9ef7841fa984be071dd72a35, type: 3}
   - <name>k__BackingField: Unagi
     <food>k__BackingField: 10
     <tag>k__BackingField: 0
@@ -202,7 +202,7 @@ MonoBehaviour:
     <isBought>k__BackingField: 0
     <level>k__BackingField: 1
     <prefab>k__BackingField: {fileID: 5773488423304670987, guid: a8fb2a1e56c1f41699379752eb62a4ce, type: 3}
-    <icon>k__BackingField: {fileID: 2800000, guid: a7ebe0897cc8b4fe9be4d6d94df071c2, type: 3}
+    <icon>k__BackingField: {fileID: 2800000, guid: 671eb2070374a4eb294a78e9033c44fb, type: 3}
   - <name>k__BackingField: CookedRice
     <food>k__BackingField: 58
     <tag>k__BackingField: 1
