+++ Assets/ScriptableObjects/InteriorObjectSO/InteriorDatabase.asset
@@ -82,14 +82,14 @@ MonoBehaviour:
     <price>k__BackingField: 1000
     <stock>k__BackingField: 0
     <Size>k__BackingField: {x: 1, y: 1}
-    <Prefab>k__BackingField: {fileID: 8881790221885300765, guid: 858fb0b7a12194f92bc869c9cc67ae8c, type: 3}
+    <Prefab>k__BackingField: {fileID: 9065045603314929499, guid: 142b17a42945b46309b36f33ac88e760, type: 3}
   - <name>k__BackingField: InductionPot
     <interior>k__BackingField: 9
     <ID>k__BackingField: 10
     <price>k__BackingField: 1000
     <stock>k__BackingField: 0
     <Size>k__BackingField: {x: 1, y: 1}
-    <Prefab>k__BackingField: {fileID: 117445688063837299, guid: 7a2712720b9ac48dbb5b730be992db29, type: 3}
+    <Prefab>k__BackingField: {fileID: 7110476475220613702, guid: 8dc84325964bc4e1fb31e5bdfdc1a9fd, type: 3}
   - <name>k__BackingField: Refridgerator
     <interior>k__BackingField: 10
     <ID>k__BackingField: 11
@@ -103,7 +103,7 @@ MonoBehaviour:
     <price>k__BackingField: 1000
     <stock>k__BackingField: 0
     <Size>k__BackingField: {x: 1, y: 1}
-    <Prefab>k__BackingField: {fileID: 8769318129579243395, guid: cccad095a94834cc5ba3729f956f0fb7, type: 3}
+    <Prefab>k__BackingField: {fileID: 6032747239418706747, guid: 1650e6ae8988c48b68631338f86c4405, type: 3}
   - <name>k__BackingField: Chair
     <interior>k__BackingField: 12
     <ID>k__BackingField: 13
