--- Assets/Scripts/KitchenInteriorWithAddandRemove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3fa5e18ad0f87428992559316ba36198
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
