--- Assets/Scripts/SimpleStuffController.cs
@@ -1,15 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI; 
-using System.Collections.Generic;
-using System.Collections; // 코루틴에서 필요한 네임스페이스
-using Yogaewonsil.Common;
-using TMPro; // TextMeshPro 사용을 위한 네임스페이스
-
-
-/// <summary>
-/// 주방테이블을 제어하는 클래스입니다.
-/// </summary>
-public class SimpleStuffController : KitchenInteriorWithAddandRemove
-{
-    // 주방테이블은 KitchenInteriorWithAddandRemove의 기본 동작을 그대로 사용합니다.
-}
