--- Assets/Scripts/SimpleStuffController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a8168eaa8690043e48ac89af8ec7a7d7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
